fixed all warnings now compiles with -Wall -Werror

+++ src/nat_check.cpp
@@ -171,8 +171,8 @@ void CNatRxManager::handle_packet_ipv4(CNatOption * option,
 }
 
 
-#define MYDP(f) if (f)  fprintf(fd," %-40s: %llu \n",#f,f)
-#define MYDP_A(f)     fprintf(fd," %-40s: %llu \n",#f,f)
+#define MYDP(f) if (f)  fprintf(fd," %-40s: %llu \n",#f,(unsigned long long)f)
+#define MYDP_A(f)     fprintf(fd," %-40s: %llu \n",#f, (unsigned long long)f)
